+++ common/v2/components/TransactionFlow/TxReceipt.scss
