+++ common/v2/components/TransactionFlow/displays/index.ts
@@ -1,6 +1,6 @@
 export * from './TransactionFeeDisplay';
 export * from './TransactionValuesDisplay';
 export { default as TransactionDetailsDisplay } from './TransactionDetailsDisplay';
-export * from './TransactionIntermediaryDisplay';
+export * from './TxIntermediaryDisplay';
 export { default as FromToAccount } from './FromToAccount';
 export { default as SwapFromToDiagram } from './SwapFromToDiagram';
