+++ common/v2/components/TransactionFlow/index.ts
@@ -1,2 +1,2 @@
 export { default as ConfirmTransaction } from './ConfirmTransaction';
-export { default as TransactionReceipt } from './TransactionReceipt';
+export { default as TxReceipt } from './TxReceipt';
