+++ common/v2/features/DeFiZap/components/ZapReceipt.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { TransactionReceipt } from 'v2/components/TransactionFlow';
+import { TxReceipt } from 'v2/components/TransactionFlow';
 import { ITxReceipt, ITxConfig, ITxType } from 'v2/types';
 import { translateRaw } from 'v2/translations';
 import { IZapConfig } from '../config';
@@ -12,14 +12,9 @@ interface Props {
   onComplete(): void;
 }
 
-export default function ZapInteractionReceipt({
-  txReceipt,
-  txConfig,
-  zapSelected,
-  onComplete
-}: Props) {
+export default function ZapReceipt({ txReceipt, txConfig, zapSelected, onComplete }: Props) {
   return (
-    <TransactionReceipt
+    <TxReceipt
       txReceipt={txReceipt}
       txConfig={txConfig}
       txType={ITxType.DEFIZAP}
