+++ common/v2/features/DeFiZap/components/index.ts
@@ -1,7 +1,7 @@
 export { default as ZapEducation } from './ZapSelection';
 export { default as ZapForm } from './ZapForm';
 export { default as ConfirmZapInteraction } from './ConfirmZapInteraction';
-export { default as ZapInteractionReceipt } from './ZapInteractionReceipt';
+export { default as ZapReceipt } from './ZapReceipt';
 export { default as DashboardZapCTA } from './DashboardZapCTA';
 export { default as zapCard } from './ZapCard';
 export { default as DetailsList } from './DetailsList';
