+++ common/v2/features/InteractWithContracts/components/InteractionReceipt.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { TransactionReceipt } from 'v2/components/TransactionFlow';
+import { TxReceipt } from 'v2/components/TransactionFlow';
 import { ITxReceipt, ITxConfig } from 'v2/types';
 import { translateRaw } from 'v2/translations';
 
@@ -14,7 +14,7 @@ export default function InteractionReceipt(props: Props) {
   const { txReceipt, txConfig, goToFirstStep } = props;
 
   return (
-    <TransactionReceipt
+    <TxReceipt
       txReceipt={txReceipt}
       txConfig={txConfig}
       completeButtonText={translateRaw('INTERACT_ANOTHER')}
